+++ resolwe_bio/processes/alignment/star.py
@@ -59,7 +59,7 @@ class AlignmentStar(Process):
     slug = "alignment-star"
     name = "STAR"
     process_type = "data:alignment:bam:star"
-    version = "5.3.2"
+    version = "5.4.2"
     category = "Align"
     scheduling_class = SchedulingClass.BATCH
     entity = {"type": "sample"}
@@ -505,6 +505,9 @@ class AlignmentStar(Process):
         stats = FileField(label="Statistics")
         species = StringField(label="Species")
         build = StringField(label="Build")
+        seqtk_downsampled = BooleanField(
+            label="The input reads were downsampled with Seqtk"
+        )
 
     def run(self, inputs, outputs):
         """Run analysis."""
@@ -572,6 +575,14 @@ class AlignmentStar(Process):
             )
         else:
             self.error("Wrong reads input type.")
+
+        if (
+            inputs.reads.type == "data:reads:fastq:single:seqtk"
+            or inputs.reads.type == "data:reads:fastq:paired:seqtk"
+        ):
+            outputs.seqtk_downsampled = True
+        else:
+            outputs.seqtk_downsampled = False
 
         if inputs.annotation:
             star_params.extend(